yuppi

+++ Kalayci.Services/Concrete/Entities/ProjectService.cs
@@ -1,6 +1,7 @@
 ﻿using Kalayci.Data.Abstract;
 using Kalayci.Data.Abstract.Entities;
 using Kalayci.Data.Concrete;
+using Kalayci.Data.Concrete.EntityFrameWork.Repositories;
 using Kalayci.Entities.Concrete;
 using Kalayci.Services.Abstract.Entities;
 using Kalayci.Shared.Data.Abstract;
@@ -15,12 +16,15 @@ namespace Kalayci.Services.Concrete.Entities
     public class ProjectService : GenericService<Project>, IProjectService
     {
         private readonly IProjectRepository _projectRepository;
+        private readonly IPersonelProjectRepository _personelProjectRepository;
         private readonly ISpoolRepository _spoolRepository;
         private readonly IUnitOfWork _unitOfWork;
         public ProjectService(IEntityRepository<Project> repository, IUnitOfWork unitOfWork,
-            IProjectRepository projectRepository, ISpoolRepository spoolRepository) : base(unitOfWork, repository)
+            IProjectRepository projectRepository, ISpoolRepository spoolRepository,
+            IPersonelProjectRepository personelProjectRepository) : base(unitOfWork, repository)
         {
-            _unitOfWork= unitOfWork;
+            _personelProjectRepository=personelProjectRepository;
+            _unitOfWork = unitOfWork;
             _spoolRepository = spoolRepository;
             _projectRepository = projectRepository;
         }
@@ -37,13 +41,21 @@ namespace Kalayci.Services.Concrete.Entities
 
         public async Task<string> RemoveProject(string ProjectId)
         {
-            ICollection<Spool> spools = await _spoolRepository.GetAllAsync(x => x.ProjectId== Convert.ToInt32(ProjectId));// 1milyon
+            ICollection<Spool> spools = await _spoolRepository.GetAllAsync(x => x.ProjectId== Convert.ToInt32(ProjectId));// 
+            ICollection<PersonelProject> personelProjects = await _personelProjectRepository.GetAllAsync(x => x.ProjectId== Convert.ToInt32(ProjectId));// 
 
 
             if (spools.Count>0)
             {
                 return $"Projeye Bağlı {spools.Count()} adet Spool Vardır önce spoolların silmelisiniz.";
             }
+
+            if (personelProjects.Count>0)
+            {
+                return $"Projeye Bağlı {personelProjects.Count()} adet Personel Vardır önce Personelleri Atamalısınız.";
+            }
+
+
             Project project = await _projectRepository.GetAsync(x => x.Id == Convert.ToInt32(ProjectId));
             await _projectRepository.DeleteAsync(project);
           await  _unitOfWork.SaveAsync();
